backend/requirements.txt
fastapi
fastmcp
python-frontmatter
pyjwt
huggingface_hub
uvicorn
httpx
pytest
pytest-asyncio
python-dotenv
requests
sqlite
